+++ controllers/MedicoController.php
@@ -1,6 +1,7 @@
 <?php
 namespace app\controllers;
 use Yii;
+use app\models\Solicitud;
 use app\models\Medico;
 use app\models\MedicoSearch;
 use app\models\PacienteSearch;
@@ -154,19 +155,14 @@ class MedicoController extends Controller {
             /*
              *   Process for ajax request
             */
-            Yii::$app
-                ->response->format = Response::FORMAT_JSON;
-            try {
+            Yii::$app->response->format = Response::FORMAT_JSON;
+            if (Solicitud::find()->where(['id_medico'=>$id])->count()>0 ){
+                return ['title' => "Eliminar medico #" . $id, 'content' => 'No se puede eliminar el medico porque esta asociado a una solicitud o más solicitudes', 'footer' => Html::button('Cerrar', ['class' => 'btn btn-default pull-left', 'data-dismiss' => "modal"]) ];
+              }
                 if ($this->findModel($id)->delete()) {
                     return ['forceClose' => true, 'forceReload' => '#crud-datatable-pjax'];
                 }
-            }
-            catch(yii\db\Exception $e) {
-                Yii::$app
-                    ->response->format = Response::FORMAT_HTML;
-                // throw new NotFoundHttpException('Error en la base de datos.',500);
-                throw new \yii\web\HttpException(500, 'No se puede eliminar el medico porque esta asociado a una solicitud', 500);
-            }
+
         }
         else {
             /*
@@ -175,35 +171,7 @@ class MedicoController extends Controller {
             return $this->redirect(['index']);
         }
     }
-    /**
-     * Delete multiple existing Medico model.
-     * For ajax request will return json object
-     * and for non-ajax request if deletion is successful, the browser will be redirected to the 'index' page.
-     * @param integer $id
-     * @return mixed
-     */
-    public function actionBulkDelete() {
-        $request = Yii::$app->request;
-        $pks = explode(',', $request->post('pks')); // Array or selected records primary keys
-        foreach ($pks as $pk) {
-            $model = $this->findModel($pk);
-            $model->delete();
-        }
-        if ($request->isAjax) {
-            /*
-             *   Process for ajax request
-            */
-            Yii::$app
-                ->response->format = Response::FORMAT_JSON;
-            return ['forceClose' => true, 'forceReload' => '#crud-datatable-pjax'];
-        }
-        else {
-            /*
-             *   Process for non-ajax request
-            */
-            return $this->redirect(['index']);
-        }
-    }
+
     /**
      * Finds the Medico model based on its primary key value.
      * If the model is not found, a 404 HTTP exception will be thrown.
