+++ controllers/TipoprofesionalController.php
@@ -2,6 +2,7 @@
 namespace app\controllers;
 use Yii;
 use app\models\Tipoprofesional;
+use app\models\Medico;
 use app\models\TipoprofesionalSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
@@ -122,52 +123,29 @@ class TipoprofesionalController extends Controller {
      * @return mixed
      */
     public function actionDelete($id) {
-        $request = Yii::$app->request;
+      $request = Yii::$app->request;
+      Yii::$app->response->format = Response::FORMAT_JSON;
+      if (Medico::find()->where(['id_tipoprofesional'=>$id])->count()>0 ){
+          return ['title' => "Eliminar tipo profesional #" . $id, 'content' => 'No se puede eliminar el tipo de profesional porque esta asociado a uno o más medicos', 'footer' => Html::button('Cerrar', ['class' => 'btn btn-default pull-left', 'data-dismiss' => "modal"]) ];
+        }
         $this->findModel($id)->delete();
-        if ($request->isAjax) {
-            /*
-             *   Process for ajax request
-            */
-            Yii::$app
-                ->response->format = Response::FORMAT_JSON;
-            return ['forceClose' => true, 'forceReload' => '#crud-datatable-pjax'];
-        }
-        else {
-            /*
-             *   Process for non-ajax request
-            */
-            return $this->redirect(['index']);
-        }
+          if ($request->isAjax) {
+              /*
+               *   Process for ajax request
+              */
+              Yii::$app
+                  ->response->format = Response::FORMAT_JSON;
+              return ['forceClose' => true, 'forceReload' => '#crud-datatable-pjax'];
+          }
+          else {
+              /*
+               *   Process for non-ajax request
+              */
+              return $this->redirect(['index']);
+          }
+
     }
-    /**
-     * Delete multiple existing Tipoprofesional model.
-     * For ajax request will return json object
-     * and for non-ajax request if deletion is successful, the browser will be redirected to the 'index' page.
-     * @param integer $id
-     * @return mixed
-     */
-    public function actionBulkDelete() {
-        $request = Yii::$app->request;
-        $pks = explode(',', $request->post('pks')); // Array or selected records primary keys
-        foreach ($pks as $pk) {
-            $model = $this->findModel($pk);
-            $model->delete();
-        }
-        if ($request->isAjax) {
-            /*
-             *   Process for ajax request
-            */
-            Yii::$app
-                ->response->format = Response::FORMAT_JSON;
-            return ['forceClose' => true, 'forceReload' => '#crud-datatable-pjax'];
-        }
-        else {
-            /*
-             *   Process for non-ajax request
-            */
-            return $this->redirect(['index']);
-        }
-    }
+
     /**
      * Finds the Tipoprofesional model based on its primary key value.
      * If the model is not found, a 404 HTTP exception will be thrown.
